fusions de tous les niveaux
