--- config.js
@@ -1,6 +1,0 @@
-const config = {
-    max_attempt : 10,
-    max_name_length : 10,
-    max_score_saved : 10
-};
-export { config };
