+++ config.mjs
@@ -1,0 +1,16 @@
+
+export default function getMaxAttempt(){
+const max_attempt = 10
+return max_attempt
+
+}
+
+export function getMaxNameLength(){
+    const max_name_length = 10
+    return max_name_length
+}
+
+export function getMaxScoreSaved(){
+    const max_score_saved = 10
+    return max_score_saved
+}
