+++ index.html
@@ -24,5 +24,5 @@
     }
   </script> -->
 </body>
-    <script type="text/javascript" src="base.js"></script>
+    <script type="module" src="base.mjs"></script>
 </html>
